+++ lib/csv_row_model.rb
@@ -24,11 +24,10 @@ require 'csv_row_model/export'
 require 'csv_row_model/export/file'
 require 'csv_row_model/export/file_model'
 
-require 'csv_row_model/validators/default_change'
-
 require 'csv_row_model/validators/number_validator'
-require 'csv_row_model/validators/boolean_format'
-require 'csv_row_model/validators/date_time_format'
-require 'csv_row_model/validators/date_format'
-require 'csv_row_model/validators/float_format'
-require 'csv_row_model/validators/integer_format'
+require 'csv_row_model/validators/default_change_validator'
+require 'csv_row_model/validators/boolean_format_validator'
+require 'csv_row_model/validators/date_time_format_validator'
+require 'csv_row_model/validators/date_format_validator'
+require 'csv_row_model/validators/float_format_validator'
+require 'csv_row_model/validators/integer_format_validator'
