+++ lib/csv_row_model/validators/boolean_format_validator.rb
